+++ packages/angular-query-persist-client/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @tanstack/angular-query-persist-client
 
+## 5.62.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/angular-query-experimental@5.90.19
+  - @tanstack/query-persist-client-core@5.91.14
+
+## 5.62.21
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/angular-query-experimental@5.90.18
+  - @tanstack/query-persist-client-core@5.91.13
+
+## 5.62.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/angular-query-experimental@5.90.17
+  - @tanstack/query-persist-client-core@5.91.12
+
 ## 5.62.19
 
 ### Patch Changes